// src/@types/master.d.ts
export type dataslistsetfieldnamesProps={
   filter:{name:string} 
}
export type masterfiednameProps={
    auth:dataslistsetfieldnamesProps
}
export type mastercateogoryfieldsProps={
    customerCategoryId:string,
    customerCategoryName:string,
}
export type masterenquiryfieldsProps={
enquireId:string,
enquireTypeName:string
}

export type masterrequirementfieldProps={
    RequirementsId:string,
    RequirementsName:string,
}
export type mastercommonfieldProps={
    name:string
}